define files on intermediate component classes
- rename python files to handler.py for consistency
- clean up autogenerated attributes in definition.yml files
- reorganize component files into separate package under components
- provide support to ComponentMeta new to copy files from base classes
- refactor component loading to separate copying files to instances
  from actually loading instance files
