+++ cirrus/cli/components/core/__init__.py
@@ -1,4 +1,3 @@
-from .. import files
 from ..base import Lambda
 from cirrus.cli.collection import Collection
 
@@ -6,11 +5,6 @@ from cirrus.cli.collection import Collection
 class CoreTask(Lambda):
     user_extendable = False
     display_type = 'Core Task'
-
-    handler = files.PythonHandler()
-    definition = files.LambdaDefinition()
-    # make this not optional once we have them
-    readme = files.Readme(optional=True)
 
 
 core_tasks = Collection(
