+++ cirrus/cli/components/core/config/process/definition.yml
@@ -1,8 +1,6 @@
 description: Consumes Process Catalogs from queue and invokes workflow
-handler: task.handler
 memorySize: 128
 timeout: 60
-module: core/process
 events:
   - sqs:
       arn: !GetAtt ProcessQueue.Arn
