+++ cirrus/cli/components/core/config/publish-test/definition.yml
@@ -1,7 +1,5 @@
 description: Test Feeder data
-handler: task.handler
 memorySize: 128
 timeout: 60
-module: core/publish-test
 events:
   - sns: ${self:provider.environment.CIRRUS_PUBLISH_TOPIC_ARN}
