+++ cirrus/cli/components/files/base.py
@@ -6,13 +6,13 @@ from pathlib import Path
 
 from cirrus.cli.exceptions import ComponentError
 from cirrus.cli.utils.console import console
-from .component import Component
 
 
 logger = logging.getLogger(__name__)
 
 
 T = TypeVar('T', bound='ComponentFile')
+Component = TypeVar('Component', bound='Component')
 class ComponentFile:
     def __init__(
         self,
@@ -51,14 +51,16 @@ class ComponentFile:
                     self._content = ''
         return self._content
 
-    def _copy_to_component(self, parent_component: Type[Component]) -> T:
-        self = copy.copy(self)
-        self.path = parent_component.path.joinpath(self.name)
-        self.parent = parent_component
+    def validate(self):
         if self.required and not self.path.is_file():
             raise ComponentError(
                 f"Cannot load {self.__class__.__name__} from '{self.path}': not a file"
             )
+
+    def _copy_to_component(self, parent_component: Type[Component]) -> T:
+        self = copy.copy(self)
+        self.path = parent_component.path.joinpath(self.name)
+        self.parent = parent_component
         return self
 
     def copy_to_component(self, component: Type[Component], name: str) -> None:
