+++ cirrus/cli/components/tasks/config/convert-to-cog/definition.yml
@@ -1,8 +1,6 @@
 description: Convert specified assets into Cloud Optimized GeoTIFFs
-handler: task.handler
 memorySize: 1024
 timeout: 900
-module: tasks/convert-to-cog
 environment:
   GDAL_DATA: /opt/share/gdal
   PROJ_LIB: /opt/share/proj
