+++ cirrus/cli/components/tasks/config/copy-assets/handler.py
